bina-backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
dateparser==1.2.0
fastapi==0.68.0
frozenlist==1.4.1
h11==0.14.0
idna==3.8
multidict==6.1.0
pydantic==1.8.2
python-binance==1.0.15
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
pytz==2024.2
regex==2024.7.24
requests==2.26.0
six==1.16.0
starlette==0.14.2
typing_extensions==4.12.2
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.15.0
websockets==9.1
yarl==1.11.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
python-multipart